refactor: group dev_worklow path
Why:
The build scripts relied on hardcoded paths to
the `dev_workflow` submodule. This made them
brittle and prone to errors if the submodule
was renamed or moved. This commit refactors the
scripts to use a variable to store the path to
the `dev_workflow` submodule. This improves
robustness and makes the scripts easier to maintain.

What:
- Updated the `init.bat` script to check for the
  existence of the dev_workflow submodule.
- Changed paths to `dev_workflow` submodule to
  use `%DEV_WORKFLOW_DIR%` environment variable.
- Updated `t_build.bat`, `t_build_maven.bat`, and
  `update-version.bat` to use the `%DEV_WORKFLOW_DIR%`
  variable for the path to the dev_workflow submodule.