fix(bat): Allow "`EOF`" to terminate version desc input
Why
The script allowed the user to enter a description for the new version.
The input was terminated by entering "`END`" on a new line.
This change allows the user to also terminate the input by entering "`EOF`".
This matches the behavior of other programs that allow for entering a block, or
lines of text, and where typing `Ctrl`+`Z` makes the system generate '`EOF`'.

What
- Added a check for "`EOF`" to terminate the input loop.
- Updated the prompt message to reflect the change.