fix(bat): update comment style to avoid errors
Why:
If you followed a :: line directly with another one, the *2nd* one
would generate a spurious "The system cannot find the drive specified."
error message and potentially execute commands inside the comment.
In the following - commented-out - example, file "out.txt" would be
created (as an empty file), and the ECHO command would execute.
	:: 1st line
	:: 2nd line > out.txt & echo HERE

More infos on https://stackoverflow.com/questions/12407800/which-comment-style-should-i-use-in-batch-files

What:
- Replaced `echo` with `REM` in `get-version.bat`, `t_build_maven.bat`, and `update-version.bat`.
- Removed unnecessary `@echo on` lines from `t_build_maven.bat`.